[Fix] Fixed pixel-precision issue in `RenderUtils#renderButton`
- Only effects MC 1.20.1 and below, since MC 1.20.2 deprecated this function with the newer sprite systems
- `RenderUtils#renderButton` now properly supports `double` values for all positioning parameters, instead of just `zLevel`

A similar change might arrive to `drawMultilineString` if there are any benefits to it...

+++ common/src/main/java/com/gitlab/cdagaming/craftpresence/utils/gui/RenderUtils.java
@@ -308,10 +308,10 @@ public class RenderUtils {
      * @param texLocation The game texture to render the button as
      */
     public static void renderButton(@Nonnull final Minecraft mc,
-                                    final int x, final int y,
-                                    final int startU, final int startV,
-                                    final int endU, final int endV,
-                                    final int width, final int height,
+                                    final double x, final double y,
+                                    final double startU, final double startV,
+                                    final double endU, final double endV,
+                                    final double width, final double height,
                                     final double zLevel,
                                     final ResourceLocation texLocation) {
         try {
